Load celdroid and attack properties from JSON
AssetManager in Model retrofitted with new AssetLoaders for celdroid and
celdroid move related properties. Created new classes/enums
CeldroidBattleMove, CeldroidDebuff, and CeldroidElement and updated
CeldroidProperties to support new method of loading. BattleScene updated
so that party is not fully initialized while on the loading screen, and
party leading celdroid has its name called out at the start of battle.
Fixed FanSelect not opening in the proper direction when the arithmetic
mean of its start and end angles is not 0.