Party lead swapping animations
Added DismissAnimation class to show a celdroid being taken off
BattleScene and SummonAnimation class to show a celdroid being spawned
on the BattleScene. A DismissAnimation followed by a SummonAnimation is
used when switching the party lead and a SummonAnimation followed by a
SummonAnimation is used at the beginning of a BattleScene.