Progress on network capabilities
DirectConnectSelectionScene opens up a popup box similar to confirmation
screens, except with a text entry field in place of buttons.  Rewrote
NioSession instantiation method to be non blocking in a similar manner
to AssetManager so that connections can be made in the game loop while
keeping interaction responsive and allowing feedback display.
DirectConnectSelectionScene will now resolve IP or DNS addresses entered
and attempt to connect using NioSession. Reduced
Constants.SOCKET_TIMEOUT. Changed color of prompt text in
ConfirmPopupScene as red should only be used for text that can be
interacted with. Added stubs for host capabilities. Simplified
DesktopGame argument parsing and added a new argument, "frame", that
executes the game with a titlebar and border so that the window can be
moved. Fixed build path order and export so that Celdroids Client
Android may use classes in Celdroids Common.